# swagger-api-v3.json
{
  "swagger": "2.0",
  "info": {
    "description": "本文档说明了每个接口返回、接收的字段信息。并且制定相应的规范作为前后端交流的渠道。",
    "version": "2.0",
    "title": "易授微服务RESTful APIs说明文档",
    "contact": {
      "name": "刘祥德",
      "url": "/",
      "email": "luvletterU@njust.edu.cn"
    }
  },
  "host": "localhost:8080",
  "basePath": "/",
...
  "paths": {
...
    "/arrangements/{arrangementId}": {
      "get": {
        "tags": [
          "teacher-arrangement-controller"
        ],
        "summary": "获取特定的教师安排表",
        "operationId": "findArrangementUsingGET_1",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "arrangementId",
            "in": "path",
            "description": "教师安排表的编号",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/TeacherArrangement"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "delete": {
        "tags": [
          "teacher-arrangement-controller"
        ],
        "summary": "deleteArrangement",
        "operationId": "deleteArrangementUsingDELETE",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "arrangementId",
            "in": "path",
            "description": "arrangementId",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/TeacherArrangement"
            }
          },
          "204": {
            "description": "No Content"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
...
    "/homeworks": {
      "get": {
        "tags": [
          "homework-controller"
        ],
        "summary": "findAllHomeworks",
        "operationId": "findAllHomeworksUsingGET",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Homework"
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "post": {
        "tags": [
          "homework-controller"
        ],
        "summary": "createHomework",
        "operationId": "createHomeworkUsingPOST",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "homework",
            "description": "homework",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Homework"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Homework"
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "put": {
        "tags": [
          "homework-controller"
        ],
        "summary": "updateHomework",
        "operationId": "updateHomeworkUsingPUT",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "homework",
            "description": "homework",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Homework"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Homework"
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
...
    "/schedules": {
      "get": {
        "tags": [
          "class-schedule-controller"
        ],
        "summary": "findAllSchedule",
        "operationId": "findAllScheduleUsingGET",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/ClassSchedule"
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "post": {
        "tags": [
          "class-schedule-controller"
        ],
        "summary": "createSchedule",
        "operationId": "createScheduleUsingPOST",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "userId",
            "in": "query",
            "description": "userId",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "name": "arrangeId",
            "in": "query",
            "description": "arrangeId",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/ClassSchedule"
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/schedules/{scheduleId}": {
      "delete": {
        "tags": [
          "class-schedule-controller"
        ],
        "summary": "deleteSchedule",
        "operationId": "deleteScheduleUsingDELETE",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "scheduleId",
            "in": "path",
            "description": "scheduleId",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/ClassSchedule"
            }
          },
          "204": {
            "description": "No Content"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
    "/schedules/{scheduleId}/arrangement": {
      "get": {
        "tags": [
          "class-schedule-controller"
        ],
        "summary": "findArrangement",
        "operationId": "findArrangementUsingGET",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "scheduleId",
            "in": "path",
            "description": "scheduleId",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/TeacherArrangement"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/schedules/{scheduleId}/doing_tests": {
      "get": {
        "tags": [
          "class-schedule-controller"
        ],
        "summary": "获取学生所有还未完成的课堂测试",
        "operationId": "findTestsIsExpectedDoneUsingGET",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "scheduleId",
            "in": "path",
            "description": "课表编号",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/OnlineClassTest"
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/schedules/{scheduleId}/done_tests": {
      "get": {
        "tags": [
          "class-schedule-controller"
        ],
        "summary": "获取所有已经被学生完成的课堂测试",
        "operationId": "findTestsIsDoneUsingGET",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "scheduleId",
            "in": "path",
            "description": "课表编号",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/OnlineClassTest"
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/schedules/{scheduleId}/homeworks": {
      "get": {
        "tags": [
          "class-schedule-controller"
        ],
        "summary": "findHomeworks",
        "operationId": "findHomeworksUsingGET",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "scheduleId",
            "in": "path",
            "description": "scheduleId",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Homework"
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "post": {
        "tags": [
          "class-schedule-controller"
        ],
        "summary": "uploadHomeWork",
        "operationId": "uploadHomeWorkUsingPOST",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "homework",
            "description": "homework",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Homework"
            }
          },
          {
            "name": "scheduleId",
            "in": "path",
            "description": "scheduleId",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Homework"
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/schedules/{scheduleId}/sign_records": {
      "get": {
        "tags": [
          "class-schedule-controller"
        ],
        "summary": "findSignRecords",
        "operationId": "findSignRecordsUsingGET",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "scheduleId",
            "in": "path",
            "description": "scheduleId",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/SignRecord"
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/schedules/{scheduleId}/student": {
      "get": {
        "tags": [
          "class-schedule-controller"
        ],
        "summary": "findStudents",
        "operationId": "findStudentsUsingGET",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "scheduleId",
            "in": "path",
            "description": "scheduleId",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Student"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/schedules/{scheduleId}/test_records": {
      "get": {
        "tags": [
          "class-schedule-controller"
        ],
        "summary": "findTestRecords",
        "operationId": "findTestRecordsUsingGET",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "scheduleId",
            "in": "path",
            "description": "scheduleId",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/TestRecord"
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/sign_records": {
      "get": {
        "tags": [
          "sign-record-controller"
        ],
        "summary": "findAllSignRecords",
        "operationId": "findAllSignRecordsUsingGET",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/SignRecord"
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "post": {
        "tags": [
          "sign-record-controller"
        ],
        "summary": "createSignRecord",
        "operationId": "createSignRecordUsingPOST",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "signRecord",
            "description": "signRecord",
            "required": true,
            "schema": {
              "$ref": "#/definitions/SignRecord"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/SignRecord"
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "put": {
        "tags": [
          "sign-record-controller"
        ],
        "summary": "updateSignRecord",
        "operationId": "updateSignRecordUsingPUT",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "signRecord",
            "description": "signRecord",
            "required": true,
            "schema": {
              "$ref": "#/definitions/SignRecord"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/SignRecord"
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
...
    "/test_records": {
      "get": {
        "tags": [
          "test-record-controller"
        ],
        "summary": "findAllTestRecords",
        "operationId": "findAllTestRecordsUsingGET",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/TestRecord"
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "put": {
        "tags": [
          "test-record-controller"
        ],
        "summary": "updateTestRecord",
        "operationId": "updateTestRecordUsingPUT",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "testRecord",
            "description": "testRecord",
            "required": true,
            "schema": {
              "$ref": "#/definitions/TestRecord"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/TestRecord"
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/test_records/{recordId}": {
      "post": {
        "tags": [
          "test-record-controller"
        ],
        "summary": "createTestRecord",
        "operationId": "createTestRecordUsingPOST",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "testRecord",
            "description": "testRecord",
            "required": true,
            "schema": {
              "$ref": "#/definitions/TestRecord"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/TestRecord"
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
...
  },
  "definitions": {
    "Attendance": {
      "type": "object",
      "properties": {
        "arrangement": {
          "$ref": "#/definitions/TeacherArrangement"
        },
        "attend": {
          "type": "boolean"
        },
        "attendanceId": {
          "type": "integer",
          "format": "int32"
        },
        "closed": {
          "type": "boolean"
        },
        "closedTime": {
          "type": "string",
          "format": "date-time"
        },
        "deadline": {
          "type": "string",
          "format": "date-time"
        },
        "establishedTime": {
          "type": "string",
          "format": "date-time"
        },
        "qRcode": {
          "$ref": "#/definitions/QRcode"
        }
      }
    },
    "ClassSchedule": {
      "type": "object",
      "properties": {
        "courseComment": {
          "$ref": "#/definitions/CourseComment"
        },
        "scheduleId": {
          "type": "integer",
          "format": "int32"
        },
        "student": {
          "$ref": "#/definitions/Student"
        },
        "teacherArrangement": {
          "$ref": "#/definitions/TeacherArrangement"
        }
      }
    },
...
    "Course": {
      "type": "object",
      "properties": {
        "courseId": {
          "type": "integer",
          "format": "int32"
        },
        "courseName": {
          "type": "string"
        },
        "credit": {
          "type": "number",
          "format": "double"
        },
        "type": {
          "type": "string"
        }
      }
    },
    "CourseComment": {
      "type": "object",
      "properties": {
        "arrangement": {
          "$ref": "#/definitions/TeacherArrangement"
        },
        "commentId": {
          "type": "integer",
          "format": "int32"
        },
        "schedule": {
          "$ref": "#/definitions/ClassSchedule"
        },
        "score": {
          "type": "number",
          "format": "float"
        }
      }
    },
...
    "Department": {
      "type": "object",
      "properties": {
        "depId": {
          "type": "integer",
          "format": "int32"
        },
        "depName": {
          "type": "string"
        }
      }
    },
    "Homework": {
      "type": "object",
      "properties": {
        "content": {
          "type": "string"
        },
        "deadline": {
          "type": "string",
          "format": "date-time"
        },
        "establishedTime": {
          "type": "string",
          "format": "date-time"
        },
        "fileName": {
          "type": "string"
        },
        "filePath": {
          "type": "string"
        },
        "homeworkId": {
          "type": "integer",
          "format": "int32"
        },
        "schedule": {
          "$ref": "#/definitions/ClassSchedule"
        },
        "size": {
          "type": "integer",
          "format": "int32"
        },
        "submitted": {
          "type": "boolean"
        },
        "title": {
          "type": "string"
        }
      }
    },
...
    "OnlineClassTest": {
      "type": "object",
      "properties": {
        "arrangement": {
          "$ref": "#/definitions/TeacherArrangement"
        },
        "deadline": {
          "type": "string",
          "format": "date-time"
        },
        "establishedTime": {
          "type": "string",
          "format": "date-time"
        },
        "questions": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Question"
          }
        },
        "testId": {
          "type": "integer",
          "format": "int32"
        }
      }
    },
    "QRcode": {
      "type": "object",
      "properties": {
        "codeId": {
          "type": "integer",
          "format": "int32"
        },
        "fileName": {
          "type": "string"
        },
        "filePath": {
          "type": "string"
        }
      }
    },
    "Question": {
      "type": "object",
      "properties": {
        "answerOrder": {
          "type": "integer",
          "format": "int32"
        },
        "content": {
          "type": "string"
        },
        "course": {
          "$ref": "#/definitions/Course"
        },
        "options": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/QuestionOption"
          }
        },
        "questionId": {
          "type": "integer",
          "format": "int32"
        }
      }
    },
    "QuestionOption": {
      "type": "object",
      "properties": {
        "content": {
          "type": "string"
        },
        "optionId": {
          "type": "integer",
          "format": "int32"
        },
        "optionOrder": {
          "type": "integer",
          "format": "int32"
        }
      }
    },
    "SignRecord": {
      "type": "object",
      "properties": {
        "attendance": {
          "$ref": "#/definitions/Attendance"
        },
        "schedule": {
          "$ref": "#/definitions/ClassSchedule"
        },
        "signId": {
          "type": "integer",
          "format": "int32"
        },
        "signTime": {
          "type": "string",
          "format": "date-time"
        }
      }
    },
    "Student": {
      "type": "object",
      "properties": {
        "department": {
          "$ref": "#/definitions/Department"
        },
        "grade": {
          "type": "string"
        },
        "majorIn": {
          "type": "string"
        },
        "password": {
          "type": "string"
        },
        "studentGender": {
          "type": "string"
        },
        "studentId": {
          "type": "string"
        },
        "studentName": {
          "type": "string"
        },
        "userId": {
          "type": "integer",
          "format": "int32"
        },
        "username": {
          "type": "string"
        }
      }
    },
    "Teacher": {
      "type": "object",
      "properties": {
        "department": {
          "$ref": "#/definitions/Department"
        },
        "password": {
          "type": "string"
        },
        "teacherGender": {
          "type": "string"
        },
        "teacherId": {
          "type": "string"
        },
        "teacherName": {
          "type": "string"
        },
        "userId": {
          "type": "integer",
          "format": "int32"
        },
        "username": {
          "type": "string"
        }
      },
      "description": "教师实体"
    },
    "TeacherArrangement": {
      "type": "object",
      "properties": {
        "arrangementId": {
          "type": "integer",
          "format": "int32"
        },
        "course": {
          "$ref": "#/definitions/Course"
        },
        "place": {
          "type": "string"
        },
        "teacher": {
          "$ref": "#/definitions/Teacher"
        }
      }
    },
    "TestRecord": {
      "type": "object",
      "properties": {
        "answer": {
          "type": "string"
        },
        "recordId": {
          "type": "integer",
          "format": "int32"
        },
        "schedule": {
          "$ref": "#/definitions/ClassSchedule"
        },
        "score": {
          "type": "number",
          "format": "float"
        },
        "test": {
          "$ref": "#/definitions/OnlineClassTest"
        }
      }
    },
...
  }
}